+++ shared/space_test.go
@@ -8,11 +8,11 @@ import (
 
 func TestSpace_Validate(t *testing.T) {
 	r := require.New(t)
-	r.EqualError(ValidateSpace(LabelGroupSize-1), fmt.Sprintf("space (%d) must be a multiple of %d", LabelGroupSize-1, LabelGroupSize))
-	r.NoError(ValidateSpace(LabelGroupSize))
-	r.EqualError(ValidateSpace(LabelGroupSize+1), fmt.Sprintf("space (%d) must be a multiple of %d", LabelGroupSize+1, LabelGroupSize))
+	r.EqualError(ValidateSpace(uint64(LabelGroupSize-1)), fmt.Sprintf("space (%d) must be a multiple of %d", LabelGroupSize-1, LabelGroupSize))
+	r.NoError(ValidateSpace(uint64(LabelGroupSize)))
+	r.EqualError(ValidateSpace(uint64(LabelGroupSize+1)), fmt.Sprintf("space (%d) must be a multiple of %d", LabelGroupSize+1, LabelGroupSize))
 
-	r.EqualError(ValidateSpace(MaxSpace-1), fmt.Sprintf("space (%d) must be a multiple of 32", MaxSpace-1))
-	r.NoError(ValidateSpace(MaxSpace))
-	r.EqualError(ValidateSpace(MaxSpace+1), fmt.Sprintf("space (%d) is greater than the supported max (%d)", MaxSpace+1, MaxSpace))
+	r.EqualError(ValidateSpace(uint64(MaxSpace-1)), fmt.Sprintf("space (%d) must be a multiple of 32", MaxSpace-1))
+	r.NoError(ValidateSpace(uint64(MaxSpace)))
+	r.EqualError(ValidateSpace(uint64(MaxSpace+1)), fmt.Sprintf("space (%d) is greater than the supported max (%d)", MaxSpace+1, MaxSpace))
 }
